refactor: remove unused component
