+++ src/pages/index.js
@@ -1,5 +1,4 @@
 export { Homepage } from "./Homepage";
-export { AllVideosFeed } from "./AllVideosFeed";
 export { PlaylistFeed } from "./PlaylistFeed";
 export { PlaylistVideos } from "./PlaylistVideos";
 export { WatchLaterFeed } from "./WatchLaterFeed";
